refactor: reorganize frontend components with folder structure and BEM styling
