+++ frontend/src/components/FileCard/FileCard.module.css
@@ -1,0 +1,125 @@
+/* ============================================================================
+   FILE CARD COMPONENT
+   ============================================================================ */
+
+.fileCard {
+    position: relative;
+    width: 100%;
+    max-width: 430px;
+    height: 230px;
+    cursor: pointer;
+    transition: transform 0.3s ease;
+}
+
+.fileCard:hover {
+    transform: scale(1.05);
+}
+
+.fileCard__background {
+    position: absolute;
+    inset: 0;
+    background: var(--brand-light);
+    border-radius: 17px;
+}
+
+.fileCard__blur {
+    position: absolute;
+    top: 8px;
+    left: 19px;
+    width: 400px;
+    height: 192px;
+    background: #f7dcd4;
+    filter: blur(14.5px);
+}
+
+.fileCard__name {
+    position: absolute;
+    bottom: 20px;
+    left: 19px;
+    right: 19px;
+    max-width: 50%;
+    font-family: 'IBM Plex Sans', sans-serif;
+    font-size: 20px;
+    font-weight: 500;
+    color: var(--brand-dark);
+    letter-spacing: 0.5px;
+    overflow: hidden;
+    text-overflow: ellipsis;
+    white-space: nowrap;
+}
+
+.fileCard__menu {
+    position: relative;
+}
+
+.fileCard__menu__button {
+    position: absolute;
+    top: 12px;
+    right: 12px;
+    width: 40px;
+    height: 40px;
+    padding: 0;
+    border: none;
+    border-radius: 8px;
+    background: var(--color-white-overlay);
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    cursor: pointer;
+    transition: all 0.2s ease;
+}
+
+.fileCard__menu__button:hover {
+    background: var(--color-primary-overlay);
+    transform: scale(1.1);
+}
+
+.fileCard__menu__button__icon {
+    width: 24px;
+    height: 24px;
+    color: var(--brand-primary);
+}
+
+.fileCard__menu__dropdown {
+    position: absolute;
+    top: 50px;
+    right: 0;
+    min-width: 160px;
+    background: var(--color-white);
+    border: 1px solid #ddd;
+    border-radius: 10px;
+    box-shadow: 0px 4px 12px rgba(0, 0, 0, 0.15);
+    overflow: hidden;
+    z-index: 50;
+}
+
+.fileCard__menu__item {
+    display: block;
+    width: 100%;
+    padding: 12px 16px;
+    border: none;
+    border-bottom: 1px solid #f0f0f0;
+    background: none;
+    text-align: left;
+    font-family: 'IBM Plex Sans', sans-serif;
+    font-size: 14px;
+    color: var(--brand-dark);
+    cursor: pointer;
+    transition: background-color 0.2s ease;
+}
+
+.fileCard__menu__item:last-child {
+    border-bottom: none;
+}
+
+.fileCard__menu__item:hover {
+    background-color: #f5f5f5;
+}
+
+.fileCard__menu__item--delete {
+    color: var(--brand-primary);
+}
+
+.fileCard__menu__item--delete:hover {
+    background-color: rgba(222, 115, 86, 0.1);
+}
