+++ frontend/src/components/LeftSidebar/index.tsx
@@ -1,8 +1,8 @@
 import { ChevronDown, Folder, Trash2, Files, Package } from "lucide-react";
 import { Link, useLocation } from "react-router-dom";
 import { useState, useEffect } from "react";
-import styles from "../styles/components.module.css";
-import { FilesService, FolderItem } from "../services/files.service";
+import styles from "./LeftSidebar.module.css";
+import { FilesService, FolderItem } from "../../services/files.service";
 
 export function LeftSidebar({ foldersExpanded, setFoldersExpanded }: { foldersExpanded: boolean; setFoldersExpanded: (expanded: boolean) => void }) {
   const location = useLocation();
@@ -30,30 +30,33 @@ export function LeftSidebar({ foldersExpanded, setFoldersExpanded }: { foldersEx
   return (
     <div className={styles.sidebar}>
       <div>
-        <Link to="/" className={styles.sidebarLogo}>
+        <Link to="/" className={styles.sidebar__logo}>
           <Package className="size-9" />
           ft_box
         </Link>
       </div>
 
-
-      <Link to="/dashboard" className={`${styles.sidebarSection} ${location.pathname === '/dashboard' ? styles.sidebarSectionActive : ''}`} style={{ top: '130px', textDecoration: 'none' }}>
-        <Files className={styles.sidebarIcon} />
-        <span className={styles.sidebarButtonLabel}>
+      <Link
+        to="/dashboard"
+        className={`${styles.sidebar__button} ${location.pathname === '/dashboard' ? styles["sidebar__button--active"] : ''}`}
+        style={{ top: '130px', textDecoration: 'none' }}
+      >
+        <Files className={styles.sidebar__button__icon} />
+        <span className={styles.sidebar__button__label}>
           All files
         </span>
       </Link>
 
-        <button
-          onClick={() => setFoldersExpanded(!foldersExpanded)}
-          className={styles.sidebarFoldersButton}
-          style={{ top: '190px' }}
-        >
-          <ChevronDown className={`${styles.sidebarChevron} ${foldersExpanded ? '' : styles.sidebarChevronRotated}`} />
-          <span className={styles.sidebarButtonLabel}>
-            Folders
-          </span>
-        </button>
+      <button
+        onClick={() => setFoldersExpanded(!foldersExpanded)}
+        className={styles.sidebar__button}
+        style={{ top: '190px' }}
+      >
+        <ChevronDown className={`${styles.sidebar__chevron} ${foldersExpanded ? '' : styles.sidebar__chevron__rotated}`} />
+        <span className={styles.sidebar__button__label}>
+          Folders
+        </span>
+      </button>
 
       {/* Dynamically render folder items */}
       {foldersExpanded && folders.length > 0 && !loading && (
@@ -61,11 +64,11 @@ export function LeftSidebar({ foldersExpanded, setFoldersExpanded }: { foldersEx
           {folders.map((folder, index) => (
             <button
               key={folder.id}
-              className={styles.sidebarFolderItem}
+              className={styles.sidebar__folder}
               style={{ top: `${250 + index * 60}px` }}
             >
-              <Folder className={styles.sidebarFolderIcon} />
-              <span className={styles.sidebarButtonLabel}>
+              <Folder className={styles.sidebar__folder__icon} />
+              <span className={styles.sidebar__button__label}>
                 {folder.name}
               </span>
             </button>
@@ -73,10 +76,13 @@ export function LeftSidebar({ foldersExpanded, setFoldersExpanded }: { foldersEx
         </>
       )}
 
-
-      <Link to="/trash" className={`${styles.sidebarTrashButton} ${location.pathname === '/trash' ? styles.sidebarTrashButtonActive : ''}`} style={{ top: foldersExpanded && folders.length > 0 ? `${250 + folders.length * 60}px` : '250px', textDecoration: 'none' }}>
-        <Trash2 className={styles.sidebarTrashIcon} />
-        <span className={styles.sidebarButtonLabel}>
+      <Link
+        to="/trash"
+        className={`${styles.sidebar__trash__button} ${location.pathname === '/trash' ? styles["sidebar__trash__button--active"] : ''}`}
+        style={{ top: foldersExpanded && folders.length > 0 ? `${250 + folders.length * 60}px` : '250px', textDecoration: 'none' }}
+      >
+        <Trash2 className={styles.sidebar__trash__button__icon} />
+        <span className={styles.sidebar__button__label}>
           Trash
         </span>
       </Link>
