+++ frontend/src/components/ProfileDropdown/ProfileDropdown.module.css
@@ -1,0 +1,141 @@
+/* ============================================================================
+   PROFILE DROPDOWN COMPONENT
+   ============================================================================ */
+
+.overlay {
+    position: fixed;
+    inset: 0;
+    z-index: 999;
+}
+
+.container {
+    position: relative;
+}
+
+.dropdown__trigger {
+    display: flex;
+    align-items: center;
+    gap: 8px;
+    padding: 8px 12px;
+    border: 1px solid #e0e0e0;
+    border-radius: 8px;
+    background: var(--color-white);
+    font-family: 'IBM Plex Sans', sans-serif;
+    font-size: 14px;
+    color: var(--brand-dark);
+    cursor: pointer;
+    transition: all 0.2s ease;
+}
+
+.dropdown__trigger:hover {
+    border-color: var(--brand-primary);
+    background: var(--color-primary-overlay);
+    color: var(--brand-primary);
+}
+
+.dropdown__trigger--active {
+    border-color: var(--brand-primary);
+    background: var(--brand-primary);
+    color: var(--color-white);
+}
+
+.dropdown__trigger__icon {
+    width: 16px;
+    height: 16px;
+    transition: transform 0.3s ease;
+}
+
+.dropdown__trigger--active .dropdown__trigger__icon {
+    transform: rotate(180deg);
+}
+
+.dropdown__menu {
+    position: absolute;
+    top: calc(100% + 8px);
+    right: 0;
+    min-width: 200px;
+    background: var(--color-white);
+    border: 1px solid #e0e0e0;
+    border-radius: 10px;
+    box-shadow: 0px 8px 24px rgba(0, 0, 0, 0.12);
+    overflow: hidden;
+    z-index: 1000;
+    animation: fadeIn 0.2s ease;
+}
+
+@keyframes fadeIn {
+    from {
+        opacity: 0;
+        transform: translateY(-8px);
+    }
+
+    to {
+        opacity: 1;
+        transform: translateY(0);
+    }
+}
+
+.dropdown__header {
+    padding: 16px;
+    border-bottom: 1px solid #f0f0f0;
+}
+
+.dropdown__header__email {
+    margin: 0;
+    font-family: 'IBM Plex Sans', sans-serif;
+    font-size: 12px;
+    font-weight: 600;
+    text-transform: uppercase;
+    color: #999;
+    letter-spacing: 0.5px;
+}
+
+.dropdown__item {
+    display: flex;
+    align-items: center;
+    gap: 12px;
+    width: 100%;
+    padding: 12px 16px;
+    border: none;
+    border-bottom: 1px solid #f0f0f0;
+    background: none;
+    text-align: left;
+    font-family: 'IBM Plex Sans', sans-serif;
+    font-size: 14px;
+    color: var(--brand-dark);
+    cursor: pointer;
+    transition: all 0.2s ease;
+    text-decoration: none;
+}
+
+.dropdown__item:last-child {
+    border-bottom: none;
+}
+
+.dropdown__item:hover {
+    background: var(--color-primary-overlay);
+    color: var(--brand-primary);
+    padding-left: 20px;
+}
+
+.dropdown__item__icon {
+    width: 18px;
+    height: 18px;
+    flex-shrink: 0;
+}
+
+.dropdown__item--danger {
+    color: #d32f2f;
+}
+
+.dropdown__item--danger:hover {
+    background: rgba(211, 47, 47, 0.1);
+    color: #c62828;
+    padding-left: 20px;
+}
+
+.dropdown__divider {
+    height: 1px;
+    background: #f0f0f0;
+    margin: 8px 0;
+}
