+++ frontend/src/pages/Dashboard/Dashboard.module.css
@@ -1,0 +1,86 @@
+/* ============================================================================
+   PAGE WRAPPER
+   ============================================================================ */
+
+.page {
+    position: relative;
+    width: 100%;
+    min-height: 100vh;
+    background-color: #f5f5f5;
+}
+
+/* ============================================================================
+   CONTENT AREA SECTION
+   ============================================================================ */
+
+.contentArea {
+    position: absolute;
+    left: 288px;
+    top: 300px;
+    right: 48px;
+}
+
+.contentAreaNoButtons {
+    position: absolute;
+    left: 288px;
+    top: 150px;
+    right: 48px;
+}
+
+.title {
+    margin-top: 0;
+    margin-bottom: 16px;
+    font-family: 'IBM Plex Sans', sans-serif;
+    font-size: 40px;
+    font-weight: 700;
+    color: var(--brand-dark);
+    letter-spacing: 0px;
+}
+
+.subtitle {
+    margin-top: 0;
+    margin-bottom: 48px;
+    font-family: 'IBM Plex Sans', sans-serif;
+    font-size: 28px;
+    font-weight: 500;
+    color: var(--brand-dark);
+    letter-spacing: 0px;
+}
+
+/* ============================================================================
+   FILE GRID SECTION
+   ============================================================================ */
+
+.fileGrid {
+    display: grid;
+    grid-template-columns: repeat(4, 1fr);
+    column-gap: 5px;
+    row-gap: 12px;
+    padding-bottom: 48px;
+}
+
+/* ============================================================================
+   RESPONSIVE BREAKPOINTS
+   ============================================================================ */
+
+@media (max-width: 1400px) {
+    .fileGrid {
+        grid-template-columns: repeat(2, 1fr);
+    }
+}
+
+@media (max-width: 900px) {
+    .fileGrid {
+        grid-template-columns: 1fr;
+    }
+
+    .searchContainer {
+        left: 200px;
+        width: 70%;
+    }
+
+    .contentArea {
+        left: 200px;
+        right: 24px;
+    }
+}
