+++ frontend/src/pages/Privacy/index.tsx
