+++ frontend/src/pages/Terms/index.tsx
