--- frontend/src/styles/dashboard.module.css
@@ -1,200 +1,0 @@
-/* ============================================================================
-   PAGE WRAPPER
-   ============================================================================ */
-
-.page {
-    position: relative;
-    width: 100%;
-    min-height: 100vh;
-    background-color: #f5f5f5;
-}
-
-/* ============================================================================
-   SEARCH BAR SECTION
-   ============================================================================ */
-
-.searchContainer {
-    position: absolute;
-    left: 284px;
-    top: 33px;
-    width: 806px;
-}
-
-.searchBox {
-    display: flex;
-    align-items: center;
-    height: 60px;
-    padding: 0 20px;
-    gap: 12px;
-    background-color: var(--color-white);
-    border: 2px solid #bbb;
-    border-radius: 12px;
-    box-shadow: 0px 2px 8px var(--color-light-shadow);
-    transition: all 0.2s ease;
-}
-
-.searchBox:focus-within {
-    border-color: #bbb;
-    box-shadow: 0px 2px 8px var(--color-light-shadow);
-}
-
-.searchIcon {
-    flex-shrink: 0;
-    width: 24px;
-    height: 24px;
-    color: #bbb;
-}
-
-.searchInput {
-    flex: 1;
-    background: transparent;
-    border: none;
-    outline: none;
-    font-family: 'IBM Plex Sans', sans-serif;
-    font-size: 18px;
-    color: #bbb;
-    letter-spacing: 0px;
-}
-
-.searchInput::placeholder {
-    color: #bbb;
-}
-
-/* ============================================================================
-   PROFILE BUTTON SECTION
-   ============================================================================ */
-
-.profileButtonContainer {
-    position: absolute;
-    right: 100px;
-    top: 33px;
-    display: flex;
-    align-items: center;
-    gap: 30px;
-}
-
-.profileButton {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    width: 60px;
-    height: 60px;
-    border: none;
-    border-radius: 50%;
-    background-color: rgba(230, 225, 224, 0.684);
-    cursor: pointer;
-    transition: all 0.2s ease;
-    box-shadow: 0px 3px 10px rgba(0, 0, 0, 0.139);
-}
-
-.profileButton:hover {
-    background-color: rgba(230, 225, 224, 0.819);
-    transform: scale(1.05);
-}
-
-.profileButtonOpen {
-    background-color: rgba(230, 225, 224, 0.819);
-}
-
-.profileIcon {
-    width: 30px;
-    height: 30px;
-    color: var(--brand-primary);
-}
-
-.settingsIconButton {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    padding: 0;
-    background: none;
-    border: none;
-    cursor: pointer;
-    transition: all 0.2s ease;
-}
-
-.settingsIconButton:hover {
-    transform: scale(1.1);
-}
-
-.settingsIconOnly {
-    width: 32px;
-    height: 32px;
-    color: var(--brand-primary);
-}
-
-/* ============================================================================
-   CONTENT AREA SECTION
-   ============================================================================ */
-
-.contentArea {
-    position: absolute;
-    left: 288px;
-    top: 300px;
-    right: 48px;
-}
-
-.contentAreaNoButtons {
-    position: absolute;
-    left: 288px;
-    top: 150px;
-    right: 48px;
-}
-
-.title {
-    margin-top: 0;
-    margin-bottom: 16px;
-    font-family: 'IBM Plex Sans', sans-serif;
-    font-size: 40px;
-    font-weight: 700;
-    color: var(--brand-dark);
-    letter-spacing: 0px;
-}
-
-.subtitle {
-    margin-top: 0;
-    margin-bottom: 48px;
-    font-family: 'IBM Plex Sans', sans-serif;
-    font-size: 28px;
-    font-weight: 500;
-    color: var(--brand-dark);
-    letter-spacing: 0px;
-}
-
-/* ============================================================================
-   FILE GRID SECTION
-   ============================================================================ */
-
-.fileGrid {
-    display: grid;
-    grid-template-columns: repeat(4, 1fr);
-    column-gap: 5px;
-    row-gap: 12px;
-    padding-bottom: 48px;
-}
-
-/* ============================================================================
-   RESPONSIVE BREAKPOINTS
-   ============================================================================ */
-
-@media (max-width: 1400px) {
-    .fileGrid {
-        grid-template-columns: repeat(2, 1fr);
-    }
-}
-
-@media (max-width: 900px) {
-    .fileGrid {
-        grid-template-columns: 1fr;
-    }
-
-    .searchContainer {
-        left: 200px;
-        width: 70%;
-    }
-
-    .contentArea {
-        left: 200px;
-        right: 24px;
-    }
-}
